% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpw_s5_alp_helper.R
\name{alp_jac}
\alias{alp_jac}
\title{Jacobian of the ALP estimating equations}
\usage{
alp_jac(beta, Xc, Xp, wts_sp, label = "One_Ref_ALP", tol_singular = 1e-07, ...)
}
\description{
Jacobian of the ALP estimating equations
}
\keyword{internal}
