% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpw_s5_alp_helper.R
\name{alp_start}
\alias{alp_start}
\title{Starting values for the coefficients of the ALP participation model}
\usage{
alp_start(Xc, Xp, wts_sp)
}
\description{
Starting values for the coefficients of the ALP participation model
}
\keyword{internal}
