% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpw_s5_clw_helper.R
\name{clw_jac}
\alias{clw_jac}
\title{Jacobian of the CLW estimating equations}
\usage{
clw_jac(beta, Xc, Xp, wts_sp, label = "One_Ref_CLW", tol_singular = 1e-07, ...)
}
\description{
Jacobian of the CLW estimating equations
}
\keyword{internal}
