% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpw_s5_clw_helper.R
\name{clw_start}
\alias{clw_start}
\title{Starting values for the coefficients of the CLW participation model}
\usage{
clw_start(Xc, Xp, wts_sp)
}
\description{
Starting values for the coefficients of the CLW participation model
}
\keyword{internal}
