% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.na.action.R
\name{na.action.pw_fit}
\alias{na.action.pw_fit}
\title{Extract NA action from a pw_fit object}
\usage{
\method{na.action}{pw_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"pw_fit"} returned by \code{\link{est_pw}}.}

\item{...}{Additional arguments (not used).}
}
\description{
Returns the \code{na.action} component recorded during the build step.
}
