% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpw_s5_cali_helper.R
\name{raking_jac}
\alias{raking_jac}
\title{Jacobian of the calibration estimating equations}
\usage{
raking_jac(
  beta,
  Xc,
  f_p,
  label = "One_Ref_Calibration",
  tol_singular = 1e-07,
  ...
)
}
\description{
Jacobian of the calibration estimating equations
}
\keyword{internal}
