% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpw_s5_cali_helper.R
\name{raking_start}
\alias{raking_start}
\title{Starting values for the coefficients of the calibration participation model}
\usage{
raking_start(Xc, f_p)
}
\description{
Starting values for the coefficients of the calibration participation model
}
\keyword{internal}
