% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.est_pw.R
\name{summary.pw_fit}
\alias{summary.pw_fit}
\title{Summarize a Pseudo-Weight Fit}
\usage{
\method{summary}{pw_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"pw_fit"}, returned by \code{\link{est_pw}}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarize a Pseudo-Weight Fit
}
